feat: mode sdk preferences
