+++ web/sdk/react/views/preferences/index.ts
@@ -1,0 +1,1 @@
+export { default as PreferencesPage } from './preferences-page';
